Addchecking email during registration

+++ Application/Empl/Commands/DeleteCommands/DeleteEmployeeCommandHandler.cs
@@ -6,7 +6,7 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Application.Empl.Commands.DeleteCommands
 {
-    public class DeleteEmplCommandHandler : IRequestHandler<DeleteEmplCommand, PersonResponse>
+    public class DeleteEmplCommandHandler : IRequestHandler<DeleteEmplCommand, Response>
     {
         private readonly IEmployeeDbContext _dbContext;
 
@@ -15,24 +15,22 @@ namespace Application.Empl.Commands.DeleteCommands
             _dbContext = dbContext;
         }
 
-        public async Task<PersonResponse> Handle(DeleteEmplCommand request, CancellationToken cancellationToken)
+        public async Task<Response> Handle(DeleteEmplCommand request, CancellationToken cancellationToken)
         {
             var entity = await _dbContext.Employees.FirstOrDefaultAsync(x => x.Id == request.EmployeeId);
             if (entity == null)
             {
-                return new PersonResponse(400, "Employee not found", false, null);
+                return new Response(400, "Employee not found", false, null);
+            }
+            if (entity.Email == "1goldyshsergei1@gmail.com")
+            {
+                return new Response(400, "This user cannot be deleted", false, null);
             }
 
             _dbContext.Employees.Remove(entity);
             await _dbContext.SaveChangesAsync(cancellationToken);
 
-            var person = new Person
-            {
-                Id = entity.Id,
-                UserName = entity.UserName
-            };
-
-            return new PersonResponse(200, "Employee deleted successfully", true, person);
+            return new Response(200, "Employee deleted successfully", true);
         }
     }
 }
